cmd/bench: add PGO sweet runs
Sweet represents PGO runs as a unique configuration (using
`name+".pgo"`), effectively a unique "toolchain". But it is really a
separate dimension. So when reporting results, we use the same
"toolchain" for both and add the new "pgo" key that reports whether PGO
was enabled or not.

CL 498335 adds backend parsing of the results and bumps the builder
timeout for the extra runtime.

Change-Id: Iea1c4fe81b7d2a35bc0cf15c1a1be87c87f2a90b
Reviewed-on: https://go-review.googlesource.com/c/benchmarks/+/498278
Reviewed-by: Michael Knyszek <mknyszek@google.com>
Run-TryBot: Michael Pratt <mpratt@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>