sweet: bump cockroachdb benchmark run timeout
Currently the benchmark has a limit of 15 minutes for each run which is
expected to take 10 minutes. This is probably not enough of a buffer
given that the benchmark is somewhat noisy and setup/teardown isn't free
either. Taking the rule of thumb that our timeout should be at least
double the actual expected time, let's bump it even more aggressively to
30 minutes.

It's also not totally clear to me that we want/need the timeout, but
let's leave it in for now. The immediate issue is that we're losing data
on the benchmark since it's occasionally failing in the 16-core
perf builder.

Change-Id: Id766b047c6bec4afae02f8e2e4c181f100c5b1b1
Reviewed-on: https://go-review.googlesource.com/c/benchmarks/+/593815
Reviewed-by: Michael Pratt <mpratt@google.com>
Auto-Submit: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>