sweet: bypass failed benchmark targets and when running the benchmark with pgo
The current implementation of sweet, when enabled with pgo, batches all the profiling
runs before running the pgo runs. When one of the target fails in the profiling runs,
the main will terminate early and no pgo runs will be tried. This CL change this
behavior to continue running the pgo runs for successfully-profiled targets.

Change-Id: I0abc1e504ade9dbc4f2999d2359040ce6134cc6e
Reviewed-on: https://go-review.googlesource.com/c/benchmarks/+/615216
Reviewed-by: Michael Pratt <mpratt@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>