cmd/bent, benchmarks/copy: replace rsync exec's with sweet's copyDir
Turns out "cp -a" is not the same as rsync (if the target
already exists, is identical, but not writeable).  This was
a problem for gomotes, and adding rsync to the gomote bundle
is a big hammer actually, so, do this instead.

This probably helps on Windows.  Definitely won't hurt.

Change-Id: Iab9be3460cd79be7358f91341b7fb4b2e1b56645
Reviewed-on: https://go-review.googlesource.com/c/benchmarks/+/617555
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>