sweet: don't build tsgo on go1.23
tsgo requires 1.24, so don't attempt to build it if the test toolchain
is older than 1.24.

To do this, we extract the version from the toolchain and check that it
is new enough, which is more difficult than it should be due to #73369.

I suspect generally there is a way to do this with go list, but 1.23
fails to even parse go.mod due to a tool directive before it can decide
it is too old to build.

Fixes golang/go#73052.
Fixes golang/go#73049.
Fixes golang/go#73047.

Change-Id: I6a6a636c495b52cff1a6196ffee79b44aaa687ba
Reviewed-on: https://go-review.googlesource.com/c/benchmarks/+/665355
Reviewed-by: Michael Knyszek <mknyszek@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Michael Pratt <mpratt@google.com>