sweet/harnesses: update the version of Cockroachdb benchmarked
This change updates the version of Cockroachdb benchmarked by sweet.
When we submitted a change to how the network is polled in CL 564197
we discovered that older versions of Cockroachdb modify the scheduler
in a way that is incompatable with the submitted change. This updates the
version of Cockroachdb to a version where that incompatibility no
longer exists.

For golang/go#73474
For golang/go#65064

Change-Id: Ifae845e025a5b64ff2cfff65e8c508c999ffbcb4
Reviewed-on: https://go-review.googlesource.com/c/benchmarks/+/668735
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Pratt <mpratt@google.com>