sweet: store and fetch assets in CIPD instead of GCS
This change moves the assets archive from GCS to CIPD. This lets us
delete a ton of code, since CIPD handles all the things we need:
storage, compression, and validation.

To use CIPD without requiring users to install the CIPD tool (at least
for fetching) this change takes a dependency on go.chromium.org/luci.
The CIPD code is very slow moving, so it's probably OK, and breakages
will be fast and loud. However, it does force an update of some of our
other dependencies.

Notably, this change also updates the versions of gopher-lua and redigo.
The former is benchmarked directly, while the latter is used on the
client side of the Tile38 benchmark. The former will likely result in a
real change in performance. The latter is unlikely to change anything at
all, since it's only the client side, and we measure server metrics
primarily. I think both of these updates are acceptable. (The gopher-lua
we depend on is quite old, so the update is welcome.)

Change-Id: Ib1673acc6aaff8d3cee6be26c52bfee305282f6f
Reviewed-on: https://go-review.googlesource.com/c/benchmarks/+/679675
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Michael Pratt <mpratt@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>