sweet/benchmarks/go-build: build runtime before running benchmark
So we don't consider time building runtime when comparing go build time.
Instead of using go build -a to rebuild everything, we'll use a new
temporary build cache directory so that nothing is cached. Then we'll
build the runtime using that cache directory to make sure that it is
cached. Then the benchmark run build, without -a will build everything
except for the runtime and its dependencies, which have already been
cached, and that time will be used in the benchmark result.

Change-Id: I2753fc8d8438be1d437f7c65077ea22760a34b54
Reviewed-on: https://go-review.googlesource.com/c/benchmarks/+/688915
Reviewed-by: Michael Matloob <matloob@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Michael Knyszek <mknyszek@google.com>